lineboterp/lineboterp_manager/requirements.txt
line-bot-sdk
flask
gunicorn
mysql-connector-python